src-tauri/python/requirements.txt
# Core AI/ML dependencies
transformers==4.36.0
torch==2.1.0
numpy==1.24.3

# Audio processing
sounddevice==0.4.6
scipy==1.11.3
librosa==0.10.1

# Text processing
nltk==3.8.1

# Utilities
tqdm==4.66.1
python-dotenv==1.0.0

# Optional: for development
pytest==7.4.3
black==23.11.0 